Refactor into package with new explore entry point
